# apidog/openapi_live.json
{"openapi":"3.1.0","info":{"title":"CRA Mock API","description":"Simplified mock backend for CRA / Expenses / CRA Tracking pages.","version":"0.1.0"},"paths":{"/api/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_api_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/me":{"get":{"tags":["auth"],"summary":"Me","operationId":"me_api_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/cra/events":{"get":{"tags":["cra"],"summary":"List Events","operationId":"list_events_api_cra_events_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"month","in":"query","required":true,"schema":{"type":"string","pattern":"^\\d{4}-(0[1-9]|1[0-2])$","title":"Month"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CraEventOut"},"title":"Response List Events Api Cra Events Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["cra"],"summary":"Create Event","operationId":"create_event_api_cra_events_post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CraEventCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CraEventOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cra/events/{event_id}":{"put":{"tags":["cra"],"summary":"Update Event","operationId":"update_event_api_cra_events__event_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"integer","title":"Event Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CraEventUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CraEventOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["cra"],"summary":"Delete Event","operationId":"delete_event_api_cra_events__event_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"integer","title":"Event Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cra/month/{month}/submit":{"post":{"tags":["cra"],"summary":"Submit Month","operationId":"submit_month_api_cra_month__month__submit_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"month","in":"path","required":true,"schema":{"type":"string","pattern":"^\\d{4}-(0[1-9]|1[0-2])$","title":"Month"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CraMonthSubmit"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CraMonthOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cra/month/{month}/signature":{"post":{"tags":["cra"],"summary":"Upload Signature","operationId":"upload_signature_api_cra_month__month__signature_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"month","in":"path","required":true,"schema":{"type":"string","pattern":"^\\d{4}-(0[1-9]|1[0-2])$","title":"Month"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_signature_api_cra_month__month__signature_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CraMonthOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cra-tracking/months":{"get":{"tags":["cra-tracking"],"summary":"List Months","operationId":"list_months_api_cra_tracking_months_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"year","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","maximum":2100,"minimum":2000},{"type":"null"}],"title":"Year"}},{"name":"month","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","maximum":12,"minimum":1},{"type":"null"}],"title":"Month"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Page_CraMonthOut_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cra-tracking/months/{month}/import-client-cra":{"post":{"tags":["cra-tracking"],"summary":"Import Client Cra","operationId":"import_client_cra_api_cra_tracking_months__month__import_client_cra_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"month","in":"path","required":true,"schema":{"type":"string","pattern":"^\\d{4}-(0[1-9]|1[0-2])$","title":"Month"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_import_client_cra_api_cra_tracking_months__month__import_client_cra_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CraMonthOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/expenses/filter":{"post":{"tags":["expenses"],"summary":"Filter Expenses","operationId":"filter_expenses_api_expenses_filter_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoteFraisFilter"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Page_NoteFraisOut_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/expenses":{"post":{"tags":["expenses"],"summary":"Create Expense","operationId":"create_expense_api_expenses_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_expense_api_expenses_post"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoteFraisOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/expenses/{expense_id}":{"put":{"tags":["expenses"],"summary":"Update Expense","operationId":"update_expense_api_expenses__expense_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"expense_id","in":"path","required":true,"schema":{"type":"integer","title":"Expense Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoteFraisUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoteFraisOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["expenses"],"summary":"Delete Expense","operationId":"delete_expense_api_expenses__expense_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"expense_id","in":"path","required":true,"schema":{"type":"integer","title":"Expense Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/healthz":{"get":{"tags":["meta"],"summary":"Healthz","operationId":"healthz_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Healthz Healthz Get"}}}}}}},"/api/enums":{"get":{"tags":["meta"],"summary":"Get Enums","operationId":"get_enums_api_enums_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Get Enums Api Enums Get"}}}}}}}},"components":{"schemas":{"Body_create_expense_api_expenses_post":{"properties":{"type":{"$ref":"#/components/schemas/ExpenseType"},"month":{"type":"string","title":"Month"},"description":{"type":"string","title":"Description","default":""},"total_amount":{"type":"number","exclusiveMinimum":0.0,"title":"Total Amount"},"billable_to_client":{"type":"boolean","title":"Billable To Client","default":false},"comment":{"type":"string","title":"Comment","default":""},"receipt":{"anyOf":[{"type":"string","contentMediaType":"application/octet-stream"},{"type":"null"}],"title":"Receipt"}},"type":"object","required":["type","month","total_amount"],"title":"Body_create_expense_api_expenses_post"},"Body_import_client_cra_api_cra_tracking_months__month__import_client_cra_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_import_client_cra_api_cra_tracking_months__month__import_client_cra_post"},"Body_login_api_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_api_auth_login_post"},"Body_upload_signature_api_cra_month__month__signature_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_signature_api_cra_month__month__signature_post"},"CraActivity":{"type":"string","enum":["Prestation","HNO","Astreinte","CP","RTT","Maladie","Sans solde","Autre"],"title":"CraActivity"},"CraCategory":{"type":"string","enum":["Absence","Travail"],"title":"CraCategory"},"CraEventCreate":{"properties":{"categorie":{"$ref":"#/components/schemas/CraCategory"},"activity":{"$ref":"#/components/schemas/CraActivity"},"start_date":{"type":"string","format":"date","title":"Start Date"},"end_date":{"type":"string","format":"date","title":"End Date"},"all_day":{"type":"boolean","title":"All Day","default":false},"nb":{"type":"number","maximum":1.0,"exclusiveMinimum":0.0,"title":"Nb","default":1.0},"description":{"type":"string","title":"Description","default":""}},"type":"object","required":["categorie","activity","start_date","end_date"],"title":"CraEventCreate"},"CraEventOut":{"properties":{"categorie":{"$ref":"#/components/schemas/CraCategory"},"activity":{"$ref":"#/components/schemas/CraActivity"},"start_date":{"type":"string","format":"date","title":"Start Date"},"end_date":{"type":"string","format":"date","title":"End Date"},"all_day":{"type":"boolean","title":"All Day","default":false},"nb":{"type":"number","maximum":1.0,"exclusiveMinimum":0.0,"title":"Nb","default":1.0},"description":{"type":"string","title":"Description","default":""},"id":{"type":"integer","title":"Id"},"user_id":{"type":"integer","title":"User Id"},"month":{"type":"string","title":"Month"}},"type":"object","required":["categorie","activity","start_date","end_date","id","user_id","month"],"title":"CraEventOut"},"CraEventUpdate":{"properties":{"categorie":{"anyOf":[{"$ref":"#/components/schemas/CraCategory"},{"type":"null"}]},"activity":{"anyOf":[{"$ref":"#/components/schemas/CraActivity"},{"type":"null"}]},"start_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Start Date"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date"},"all_day":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"All Day"},"nb":{"anyOf":[{"type":"number","maximum":1.0,"exclusiveMinimum":0.0},{"type":"null"}],"title":"Nb"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","title":"CraEventUpdate"},"CraMonthOut":{"properties":{"id":{"type":"integer","title":"Id"},"user_id":{"type":"integer","title":"User Id"},"month":{"type":"string","title":"Month"},"status":{"$ref":"#/components/schemas/Status"},"description_tasks":{"type":"string","title":"Description Tasks"},"reserve_use_eur":{"type":"number","title":"Reserve Use Eur"},"reserve_use_days":{"type":"number","title":"Reserve Use Days"},"reserve_save_eur":{"type":"number","title":"Reserve Save Eur"},"reserve_save_days":{"type":"number","title":"Reserve Save Days"},"signature_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Signature Path"},"client_cra_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Cra Path"},"submitted_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Submitted At"},"validated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Validated At"}},"type":"object","required":["id","user_id","month","status","description_tasks","reserve_use_eur","reserve_use_days","reserve_save_eur","reserve_save_days","signature_path","client_cra_path","submitted_at","validated_at"],"title":"CraMonthOut"},"CraMonthSubmit":{"properties":{"description_tasks":{"type":"string","title":"Description Tasks","default":""},"reserve_use_eur":{"type":"number","title":"Reserve Use Eur","default":0.0},"reserve_use_days":{"type":"number","title":"Reserve Use Days","default":0.0},"reserve_save_eur":{"type":"number","title":"Reserve Save Eur","default":0.0},"reserve_save_days":{"type":"number","title":"Reserve Save Days","default":0.0}},"type":"object","title":"CraMonthSubmit"},"ExpenseType":{"type":"string","enum":["Indemnités kilométriques","Autre","Restaurant","Materiel","Titre de transport","Telephonie","Teletravail"],"title":"ExpenseType"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"NoteFraisFilter":{"properties":{"year":{"type":"integer","title":"Year"},"month":{"anyOf":[{"type":"integer","maximum":12.0,"minimum":1.0},{"type":"null"}],"title":"Month"},"status":{"anyOf":[{"$ref":"#/components/schemas/Status"},{"type":"null"}]},"type":{"anyOf":[{"$ref":"#/components/schemas/ExpenseType"},{"type":"null"}]},"billable":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Billable"},"page":{"type":"integer","minimum":1.0,"title":"Page","default":1},"limit":{"type":"integer","maximum":100.0,"minimum":1.0,"title":"Limit","default":10}},"type":"object","required":["year"],"title":"NoteFraisFilter"},"NoteFraisOut":{"properties":{"id":{"type":"integer","title":"Id"},"user_id":{"type":"integer","title":"User Id"},"month":{"type":"string","title":"Month"},"type":{"$ref":"#/components/schemas/ExpenseType"},"description":{"type":"string","title":"Description"},"total_amount":{"type":"number","title":"Total Amount"},"billable_to_client":{"type":"boolean","title":"Billable To Client"},"comment":{"type":"string","title":"Comment"},"receipt_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Receipt Path"},"status":{"$ref":"#/components/schemas/Status"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","user_id","month","type","description","total_amount","billable_to_client","comment","receipt_path","status","created_at"],"title":"NoteFraisOut"},"NoteFraisUpdate":{"properties":{"type":{"anyOf":[{"$ref":"#/components/schemas/ExpenseType"},{"type":"null"}]},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"total_amount":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"null"}],"title":"Total Amount"},"billable_to_client":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Billable To Client"},"comment":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Comment"}},"type":"object","title":"NoteFraisUpdate"},"Page_CraMonthOut_":{"properties":{"items":{"items":{"$ref":"#/components/schemas/CraMonthOut"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"},"page":{"type":"integer","title":"Page"},"pages":{"type":"integer","title":"Pages"},"limit":{"type":"integer","title":"Limit"}},"type":"object","required":["items","total","page","pages","limit"],"title":"Page[CraMonthOut]"},"Page_NoteFraisOut_":{"properties":{"items":{"items":{"$ref":"#/components/schemas/NoteFraisOut"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"},"page":{"type":"integer","title":"Page"},"pages":{"type":"integer","title":"Pages"},"limit":{"type":"integer","title":"Limit"}},"type":"object","required":["items","total","page","pages","limit"],"title":"Page[NoteFraisOut]"},"Status":{"type":"string","enum":["Draft","Pending","Verified","Validated","Rejected"],"title":"Status"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"TokenResponse"},"UserOut":{"properties":{"id":{"type":"integer","title":"Id"},"email":{"type":"string","title":"Email"},"full_name":{"type":"string","title":"Full Name"},"role":{"$ref":"#/components/schemas/UserRole"}},"type":"object","required":["id","email","full_name","role"],"title":"UserOut"},"UserRole":{"type":"string","enum":["member","admin","rh"],"title":"UserRole"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/auth/login"}}}}}}